backend/requirements.txt
fastapi[standard]
uvicorn
psycopg2-binary
python-dotenv
pytest
pytest-cov
pytest-mock